code clean up + non-admin restrictions for client

+++ client/src/services/BookService.js
@@ -1,4 +1,5 @@
 import { Api, setBearer } from '@/services/Api'
+import { isAdmin } from './AuthService'
 
 export function fetchBooks () {
   setBearer()
@@ -6,16 +7,25 @@ export function fetchBooks () {
 }
 
 export function createBook (bookData) {
+  if (!isAdmin()) {
+    throw new Error('You do not have sufficient permission.')
+  }
   setBearer()
   return Api().post('books', bookData)
 }
 
 export function editBook (bookId, bookData) {
+  if (!isAdmin()) {
+    throw new Error('You do not have sufficient permission.')
+  }
   setBearer()
   return Api().put(`books/${bookId}`, bookData)
 }
 
 export function deleteBook (bookId) {
+  if (!isAdmin()) {
+    throw new Error('You do not have sufficient permission.')
+  }
   setBearer()
   return Api().delete(`books/${bookId}`)
 }
